+++ components/frontend/src/subject/SubjectTableRow.jsx
@@ -13,7 +13,7 @@ import { Overrun } from "../measurement/Overrun"
 import { StatusIcon } from "../measurement/StatusIcon"
 import { TimeLeft } from "../measurement/TimeLeft"
 import { TrendSparkline } from "../measurement/TrendSparkline"
-import { MetricDetails } from "../metric/MetricDetails"
+import { MetricDetails, TREND_GRAPH_TAB_INDEX } from "../metric/MetricDetails"
 import { measurementOnDate } from "../report/report_utils"
 import {
     dataModelPropType,
@@ -331,8 +331,22 @@ export function SubjectTableRow({
                 />
             )}
             {!columnsToHide.includes("trend") && (
-                <TableCell sx={{ width: "150px" }}>
-                    <TrendSparkline measurements={metric.recent_measurements} reportDate={reportDate} scale={scale} />
+                <TableCell sx={{ padding: 0, width: "150px" }}>
+                    <TrendSparkline
+                        measurements={metric.recent_measurements}
+                        onClick={() => {
+                            if (
+                                settings.expandedItems.includes(metricUuid) &&
+                                settings.expandedItems.getItem(metricUuid) === TREND_GRAPH_TAB_INDEX
+                            ) {
+                                settings.expandedItems.deleteItem(metricUuid)
+                            } else {
+                                settings.expandedItems.setItem(metricUuid, TREND_GRAPH_TAB_INDEX)
+                            }
+                        }}
+                        reportDate={reportDate}
+                        scale={scale}
+                    />
                 </TableCell>
             )}
             {!columnsToHide.includes("status") && (